+++ test/noahtheduke/splint/re_frame_test.clj
@@ -23,7 +23,7 @@
     naming/conventional-aliases 2
     naming/predicate 1
     naming/single-segment-namespace 1
-    performance/assoc-many 18
+    performance/assoc-many 2
     performance/avoid-satisfies 2
     style/eq-false 1
     style/eq-nil 2
@@ -46,4 +46,4 @@
              :diagnostics
              (group-by :rule-name)
              (#(update-vals % count)))))
-    (expect 75 (count (:diagnostics results)))))
+    (expect 59 (count (:diagnostics results)))))