+++ test/noahtheduke/splint/rules/performance/assoc_many_test.clj
@@ -19,4 +19,29 @@
                 (assoc :k2 2)
                 (assoc :k3 3))}]
     "(assoc m :k1 1 :k2 2 :k3 3)"
+    (config))
+  (expect-match
+    [{:rule-name 'performance/assoc-many
+      :form '(-> m (assoc :k1 1 :k2 2 :k3 3))
+      :message "Faster to call assoc multiple times."
+      :alt '(-> m
+                (assoc :k1 1)
+                (assoc :k2 2)
+                (assoc :k3 3))}]
+    "(-> m (assoc :k1 1 :k2 2 :k3 3))"
+    (config))
+  (expect-match
+    [{:rule-name 'performance/assoc-many
+      :form '(-> m (assoc :k1 1) (assoc :k2 2) (assoc :k3 3 :k4 4))
+      :message "Faster to call assoc multiple times."
+      :alt '(-> m
+                (assoc :k1 1)
+                (assoc :k2 2)
+                (assoc :k3 3)
+                (assoc :k4 4))}]
+    "(-> m (assoc :k1 1) (assoc :k2 2) (assoc :k3 3 :k4 4))"
+    (config))
+  (expect-match
+    nil
+    "(assoc m :k1 1)"
     (config)))